moved decomposition tests to a single test file, added in a notimplemented exception in case a tensorflow object is used as a decomposition argument
